Fix ReadOnlyBoardExtensions.TryFindShip
When the square at the given board index is SquareContent.HitShip and the index is located at the border of the board, the method returned ShipFindingResult.CompleteShip in some circumstances.
In order to fix this, the method now returns ShipFindingResult.PartialResult when only a single square of the ship is revealed.